requirements.txt
aiosqlite==0.17.0
amqp==5.2.0
annotated-types==0.6.0
asgiref==3.8.1
asttokens==2.4.1
asyncpg==0.29.0
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.5.0
decorator==5.1.1
Django==5.0.4
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
executing==2.0.1
inflection==0.5.1
ipython==8.23.0
iso8601==1.1.0
jedi==0.19.1
kombu==5.3.7
matplotlib-inline==0.1.7
packaging==24.0
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.17.2
PyJWT==2.8.0
pypika-tortoise==0.1.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
setuptools==69.5.1
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
tortoise-orm==0.20.1
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
